requirements.txt
playwright>=1.40.0
supabase>=2.9.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
